% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{attribution_for_datasets}
\alias{attribution_for_datasets}
\title{Get attribution for datasets}
\usage{
attribution_for_datasets(datasets)
}
\arguments{
\item{datasets}{Vector of dataset identifiers}
}
\value{
Returns a string to be used as attribution for the given datasets.
}
\description{
Get attribution for datasets
}
\examples{

# Attribution string for the 2006 and 2016 census datasets
attribution_for_datasets(c('CA06','CA16'))
}
