% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{cancensus.as_region_list}
\alias{cancensus.as_region_list}
\title{Convert a (suitably filtered) data frame from
\code{\link{cancensus.list_regions}} to a list suitable for passing to
\code{\link{cancensus.load}}.}
\usage{
cancensus.as_region_list(tbl)
}
\arguments{
\item{tbl}{A data frame, suitably filtered, as returned by
\code{\link{cancensus.list_regions}}.}
}
\description{
Convert a (suitably filtered) data frame from
\code{\link{cancensus.list_regions}} to a list suitable for passing to
\code{\link{cancensus.load}}.
}
\examples{

library(dplyr, warn.conflicts = FALSE)

# Query the CensusMapper API for the total occupied dwellings
# of 20 random Census Subdivisions, in Census 2016.
regions <- cancensus.list_regions("CA16") \%>\%
  dplyr::filter(level == "CSD") \%>\%
  dplyr::sample_n(20) \%>\%
  cancensus.as_region_list()

occupied <- cancensus.load_data("CA16", regions = regions,
                                vectors = c("v_CA16_408"),
                                level = "Regions")

}
