% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{cancensus.handle_status_code}
\alias{cancensus.handle_status_code}
\title{Internal function to handle unfavourable http responses}
\usage{
cancensus.handle_status_code(response, path)
}
\description{
Internal function to handle unfavourable http responses
}
