% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{cancensus.labels}
\alias{cancensus.labels}
\title{Return Census variable names and labels as a tidy data frame}
\usage{
cancensus.labels(dat)
}
\value{
A data frame with a column \code{variable} containing the truncated
variable name, and a column \code{label} describing it.
}
\description{
Return Census variable names and labels as a tidy data frame
}
