% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{cancensus.list_vectors}
\alias{cancensus.list_vectors}
\title{Query the CensusMapper API for available vectors for a given dataset.}
\usage{
cancensus.list_vectors(dataset)
}
\arguments{
\item{dataset}{The dataset to query for available vectors, e.g.
\code{"CA16"}.}
}
\description{
Query the CensusMapper API for available vectors for a given dataset.
}
