% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{cancensus.load}
\alias{cancensus.load}
\title{CensusMapper API access}
\usage{
cancensus.load(dataset, level, regions, vectors = c(), geo = TRUE,
  format = "sf", use_cache = TRUE)
}
\arguments{
\item{dataset}{A CensusMapper dataset identifier.}

\item{level}{A geographic aggregation level for downloading data, e.g. CSD, CT, DA.}

\item{regions}{A json hash describing the geographic regions.}

\item{vectors}{An R vector containing the CensusMapper variable names of the census variables to download. If no vectors are specified only geographic data will get downloaded.}

\item{geo}{If set to TRUE, the function will also return the geographic data.}

\item{format}{Choose whether you want to use the sf or sp spatial format. Using sf will return a dataframe with a field for sf geometry, while using sp will return a SpatialPolygonsDataFrame object. Assumes sf as default but can be overwritten by selecting format = "sp".}

\item{use_cache}{If set to TRUE (the default) data will be read from the local cache if available.}
}
\description{
This function allows convenient access to the CensusMapper API
}
\details{
Got to https://censusmapper.ca/api to select variables and geographic regions
you want to download data for

An API key is required to use this function. Either set the API key on censusmapper
api_key='<your API key>'
or as environment variable
sys.setenv(CM_API_KEY='<your API key>')
}
\examples{
census_data <- cancensus.load(dataset='CA16', regions='{"CMA":["59933"]}', vectors=c("v_CA16_408","v_CA16_409","v_CA16_410"), level='CSD', geo=TRUE)
}
\keyword{api}
\keyword{canada}
\keyword{census}
\keyword{data}
