% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{cancensus.load_data}
\alias{cancensus.load_data}
\title{Convenience function to load only census data and no geographies.}
\usage{
cancensus.load_data(dataset, level, regions, vectors = c(),
  use_cache = TRUE, api_key = getOption("cancensus.api_key"))
}
\arguments{
\item{dataset}{A CensusMapper dataset identifier.}

\item{level}{A geographic aggregation level for downloading data, e.g. CSD, CT, DA.}

\item{regions}{A json hash describing the geographic regions.}

\item{vectors}{An R vector containing the CensusMapper variable names of the census variables to download. If no vectors are specified only geographic data will get downloaded.}

\item{use_cache}{If set to TRUE (the default) data will be read from the local cache if available.}

\item{geo}{If set to TRUE, the function will also return the geographic data.}
}
\description{
Convenience function to load only census data and no geographies.
}
\examples{
census_data <- cancensus.load_data(dataset='CA16', regions='{"CMA":["59933"]}', vectors=c("v_CA16_408","v_CA16_409","v_CA16_410"), level='CSD')
}
\keyword{api}
\keyword{canada}
\keyword{census}
\keyword{data}
