% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{cancensus.load_geo}
\alias{cancensus.load_geo}
\title{Convenience function to load only census data and no geographies.}
\usage{
cancensus.load_geo(dataset, level, regions, use_cache = TRUE)
}
\arguments{
\item{dataset}{A CensusMapper dataset identifier.}

\item{level}{A geographic aggregation level for downloading data, e.g. CSD, CT, DA.}

\item{regions}{A json hash describing the geographic regions.}

\item{use_cache}{If set to TRUE (the default) data will be read from the local cache if available.}

\item{geo}{If set to TRUE, the function will also return the geographic data.}
}
\description{
Convenience function to load only census data and no geographies.
}
\examples{
census_data <- cancensus.load_geo(dataset='CA16', regions='{"CMA":["59933"]}', level='CSD')
}
\keyword{api}
\keyword{canada}
\keyword{census}
\keyword{data}
