% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{cancensus.set_api_key}
\alias{cancensus.set_api_key}
\title{Convenience function to set the api key for the current session.}
\usage{
cancensus.set_api_key(api_key)
}
\arguments{
\item{api_key}{Your CensusMapper API key.}
}
\description{
Convenience function to set the api key for the current session.
}
\examples{
cancensus.set_api_key('CensusMapper_2e24662e6dde22b46d5a316e81bebddf')
}
