% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{census_vectors}
\alias{census_vectors}
\title{Return Census variable names and labels as a tidy data frame}
\usage{
census_vectors(x)
}
\arguments{
\item{x}{A data frame, \code{sp} or \code{sf} object returned from
\code{get_census} or similar.}
}
\value{
A data frame with a column \code{variable} containing the truncated
variable name, and a column \code{label} describing it.
}
\description{
Return Census variable names and labels as a tidy data frame
}
