% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{CODES_TABLE}
\alias{CODES_TABLE}
\title{A dataset with code table summaries for census data}
\description{
A dataset with code table summaries for census data
}
\references{
\url{https://www12.statcan.gc.ca/census-recensement/2021/geo/ref/domain-domaine/index2021-eng.cfm?lang=e&id=CSDtype}, \url{https://www12.statcan.gc.ca/census-recensement/2021/geo/ref/domain-domaine/index2021-eng.cfm?lang=e&id=CDtype}
}
\author{
derived from StatCan definitions
}
\keyword{data}
