% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{COV_SKYTRAIN_STATIONS}
\alias{COV_SKYTRAIN_STATIONS}
\title{A dataset City of Vancouver skytrain station locations}
\description{
A dataset City of Vancouver skytrain station locations
}
\references{
\url{https://opendata.vancouver.ca/explore/dataset/rapid-transit-stations/information/}
}
\author{
City of Vancouver Open Data
}
\keyword{data}
