% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{dataset_attribution}
\alias{dataset_attribution}
\title{Get attribution for datasets}
\usage{
dataset_attribution(datasets)
}
\arguments{
\item{datasets}{Vector of dataset identifiers}
}
\value{
Returns a string to be used as attribution for the given datasets.
}
\description{
Get attribution for datasets
}
\examples{

\dontrun{
# Attribution string for the 2006 and 2016 census datasets
dataset_attribution(c('CA06','CA16'))
}
}
