% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{get_census_data}
\alias{get_census_data}
\title{Convenience function to load only census data and no geographies.}
\usage{
get_census_data(dataset, level, regions, vectors, ...)
}
\arguments{
\item{dataset}{A CensusMapper dataset identifier.}

\item{level}{The census aggregation level to retrieve. One of \code{"Regions"}, \code{"PR"}, \code{"CMA"}, \code{"CD"}, \code{"CSD"}, \code{"CT"} or \code{"DA"}.}

\item{regions}{A named list of census regions to retrieve. Names must be valid census aggregation levels.}

\item{vectors}{An R vector containing the CensusMapper variable names of the census variables to download. If no vectors are specified only geographic data will get downloaded.}

\item{...}{Further arguments passed on to \code{\link[cancensus]{get_census}}.}
}
\description{
Convenience function to load only census data and no geographies.
}
\examples{
census_data <- get_census_data(dataset='CA16', regions=list(CMA:"59933"),
                               vectors=c("v_CA16_408","v_CA16_409","v_CA16_410"),
                               level='CSD')
}
\keyword{api}
\keyword{canada}
\keyword{census}
\keyword{data}
