% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_suite.R
\name{get_geographic_attributes}
\alias{get_geographic_attributes}
\title{Read the Geographic Attributes File}
\usage{
get_geographic_attributes(census_year = "2016", refresh = FALSE)
}
\arguments{
\item{census_year}{census year to get the data for, right now only 2006, 2011 and 2016 are supported}

\item{refresh}{(logical) refresh the cache if true}
}
\value{
tibble with the relationship data
}
\description{
Reads the Geographies Attributes File for the given census year. The table contains
the information on how all the geographic levels are related for each area, and population, dwelling and household counts.
A reference guide is available
at https://www150.statcan.gc.ca/n1/en/catalogue/92-151-G2021001
}
\examples{
# list add the cached census data
get_geographic_attributes("2016")

}
