% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_suite.R
\name{get_geography_relationship}
\alias{get_geography_relationship}
\title{Read the Dissemination Geographies Relationship File}
\usage{
get_geography_relationship(census_year = "2021", refresh = FALSE)
}
\arguments{
\item{census_year}{census year to get the data for, right now only 2021 is supported}

\item{refresh}{(logical) refresh the cache if true}
}
\value{
tibble with the relationship data
}
\description{
Reads the Dissemination Geographies Relationship File for the given census year. The table contains
the information on how all the geographic levels are related for each area. A reference guide is available
at https://www150.statcan.gc.ca/n1/en/catalogue/982600032021001
}
\examples{
# list add the cached census data
get_geography_relationship("2021")

}
