% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalled_data.R
\name{get_recalled_database}
\alias{get_recalled_database}
\title{Get recalled data}
\usage{
get_recalled_database()
}
\description{
Grabs recall data from server if needed and returns list of recalled data
}
\examples{
\dontrun{
check_recalled_data()
}
}
\keyword{internal}
