% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus_cache 2.R, R/cancensus_cache.R
\name{list_cancensus_cache}
\alias{list_cancensus_cache}
\title{List cached files}
\usage{
list_cancensus_cache()

list_cancensus_cache()
}
\value{
tibble with metadata on cached data

tibble with metadata on cached data
}
\description{
Lists all cached data and metadata if available

Lists all cached data and metadata if available
}
\examples{
# list add the cached census data
list_cancensus_cache()

# list add the cached census data
list_cancensus_cache()

}
