% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{list_census_datasets}
\alias{list_census_datasets}
\title{Query the CensusMapper API for available datasets.}
\usage{
list_census_datasets(use_cache = FALSE, quiet = FALSE)
}
\arguments{
\item{use_cache}{If set to TRUE, data will be read from a local cache, if
available. If set to FALSE (the default), query the API for the data, and
refresh the local cache with the result.}

\item{quiet}{When TRUE, suppress messages and warnings.}
}
\value{
Returns a data frame with a column \code{dataset} containing the code for the
dataset, and a column \code{description} describing it.
}
\description{
Query the CensusMapper API for available datasets.
}
\examples{

# List available datasets in CensusMapper
list_census_datasets()
}
