% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{list_datasets}
\alias{list_datasets}
\title{Query the CensusMapper API for available datasets.}
\usage{
list_datasets()
}
\value{
A data frame with a column \code{dataset} containing the code for the
dataset, and a column \code{description} describing it.
}
\description{
Query the CensusMapper API for available datasets.
}
