% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalled_data.R
\name{list_recalled_cached_data}
\alias{list_recalled_cached_data}
\title{Lists locally cached data that has been recalled}
\usage{
list_recalled_cached_data()
}
\description{
Checks the local cached database for recalled data and lists all recalled cached entries
}
\examples{
\dontrun{
list_recalled_cached_data()
}
}
