% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_vectors.R
\name{parent_census_vectors}
\alias{parent_census_vectors}
\title{List all parent variables from vector hierarchical based on a (sub-)list of census
variables returned by
\code{list_census_vectors} or \code{search_census_vectors}.}
\usage{
parent_census_vectors(vector_list)
}
\arguments{
\item{vector_list}{The list of vectors to be used}
}
\description{
List all parent variables from vector hierarchical based on a (sub-)list of census
variables returned by
\code{list_census_vectors} or \code{search_census_vectors}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

list_census_vectors("CA16") \%>\%
  filter(vector == "v_CA16_2519") \%>\%
  parent_census_vectors()

# Query parent vectors directly using vector identifier
parent_census_vectors("v_CA16_2519")
}
