% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus_cache 2.R, R/cancensus_cache.R
\name{remove_from_cancensus_cache}
\alias{remove_from_cancensus_cache}
\title{Remove cached files}
\usage{
remove_from_cancensus_cache(paths)

remove_from_cancensus_cache(paths)
}
\arguments{
\item{paths}{list of paths to remove, as returned by the path column in `list_cancensus_cache`}
}
\value{
freed-up disk space

freed-up disk space
}
\description{
Remove cached data for paths given

Remove cached data for paths given
}
\examples{
\dontrun{
# remove the first cached dataset
cache_data <- list_cancensus_cache()

remove_from_cancensus_cache(cache_data$path[1])
}
\dontrun{
# remove the first cached dataset
cache_data <- list_cancensus_cache()

remove_from_cancensus_cache(cache_data$path[1])
}
}
