% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalled_data.R
\name{remove_recalled_cached_data}
\alias{remove_recalled_cached_data}
\title{Remove recalled data from local cache}
\usage{
remove_recalled_cached_data()
}
\value{
Storage size of removed locally cached data that got freed up in number of bytes.
}
\description{
Checks the local cached database for recalled data and removes cached data that has been recalled
}
\examples{
\dontrun{
remove_recalled_cached_data()
}
}
