% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{search_census_vectors}
\alias{search_census_vectors}
\title{Query the CensusMapper API for vectors with descriptions matching a searchterm.}
\usage{
search_census_vectors(searchterm, dataset)
}
\arguments{
\item{searchterm}{The term to search for e.g. \code{"Ojibway"}.
Search terms are case insensitive. If unable to find a given search term,
this function will suggest the correct spelling to use when possible.}

\item{dataset}{The dataset to query for available vectors, e.g.
\code{"CA16"}.}
}
\description{
Query the CensusMapper API for vectors with descriptions matching a searchterm.
}
\examples{
search_census_vectors('Ojibway', 'CA16')

# This will return a warning that no match was found, but will suggest similar terms.
search_census_vectors('Ojibwe', 'CA16')
}
