% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_settings.R
\name{show_cancensus_cache_path}
\alias{show_cancensus_cache_path}
\title{View saved cache directory path}
\usage{
show_cancensus_cache_path()
}
\description{
View saved cache path'
}
\examples{
show_cancensus_cache_path()
}
