% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{cancensus.load}
\alias{cancensus.load}
\title{CensusMapper API access}
\usage{
cancensus.load(dataset, level, regions, vectors = c(), geo_format = "sf",
  labels = "detailed", use_cache = TRUE,
  api_key = getOption("cancensus.api_key"))
}
\arguments{
\item{dataset}{A CensusMapper dataset identifier.}

\item{level}{The census aggregation level to retrieve. One of \code{"Regions"}, \code{"PR"}, \code{"CMA"}, \code{"CD"}, \code{"CSD"}, \code{"CT"} or \code{"DA"}.}

\item{regions}{A named list of census regions to retrieve. Names must be valid census aggregation levels.}

\item{vectors}{An R vector containing the CensusMapper variable names of the census variables to download. If no vectors are specified only geographic data will get downloaded.}

\item{geo_format}{One of \code{"sf"} to return an \code{\link[sf]{sf}} object (the default; requires the \code{sf} package), \code{"sp"} to return a \code{\link[sp]{SpatialPolygonsDataFrame}} object (requires the \code{rgdal} package), or \code{NA} to append no geographical information to the returned data at all.}

\item{labels}{Set to "detailed" by default, but truncated Census variable names can be selected by setting labels = "short". Use cancensensus.labels() to return variable label information.}

\item{use_cache}{If set to TRUE (the default) data will be read from the local cache if available.}

\item{api_key}{An API key for the CensusMapper API. Defaults to \code{options()} and then the \code{CM_API_KEY} environment variable.}
}
\description{
This function allows convenient access to the CensusMapper API
}
\details{
Got to https://censusmapper.ca/api to select variables and geographic regions
you want to download data for

An API key is required to use this function. Either set the API key on censusmapper
api_key='<your API key>'
or as environment variable
sys.setenv(CM_API_KEY='<your API key>')
}
\examples{
# Load data without associated geographical spatial data
census_data <- cancensus.load(dataset='CA16', regions=list(CMA="59933"),
                              vectors=c("v_CA16_408","v_CA16_409","v_CA16_410"),
                              level='CSD', geo_format = NA)

# Load data with associated geographical spatial data using the sf standard
census_data <- cancensus.load(dataset='CA16', regions=list(CMA="59933"),
                              vectors=c("v_CA16_408","v_CA16_409","v_CA16_410"),
                              level='CSD', geo_format = "sf")

# Load data with geography and truncated variable names
census_data <- cancensus.load(dataset='CA16', regions=list(CMA="59933"),
                              vectors=c("v_CA16_408","v_CA16_409","v_CA16_410"),
                              level='CSD', geo_format = "sf", labels="short")

# Get details for truncated variables
cancensus.labels(census_data)
}
\keyword{api}
\keyword{canada}
\keyword{census}
\keyword{data}
