% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{cancensus.load_geo}
\alias{cancensus.load_geo}
\title{Convenience function to load only census geography without data.}
\usage{
cancensus.load_geo(dataset, level, regions, format, use_cache = TRUE)
}
\arguments{
\item{dataset}{A CensusMapper dataset identifier.}

\item{level}{A geographic aggregation level for downloading data, e.g. CSD, CT, DA.}

\item{regions}{A json hash describing the geographic regions.}

\item{format}{Choose whether you want to use the sf or sp spatial format. Assumes sf as default but can be overwritten by selecting format = "sp".}

\item{use_cache}{If set to TRUE (the default) data will be read from the local cache if available.}
}
\description{
Convenience function to load only census geography without data.
}
\examples{
census_data <- cancensus.load_geo(dataset='CA16', regions='{"CMA":["59933"]}', level='CSD')
}
\keyword{api}
\keyword{canada}
\keyword{census}
\keyword{data}
