% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{cancensus.load_geo}
\alias{cancensus.load_geo}
\title{Convenience function to load only census geography without data.}
\usage{
cancensus.load_geo(dataset, level, regions, geo_format = "sf",
  labels = "detailed", use_cache = TRUE,
  api_key = getOption("cancensus.api_key"))
}
\arguments{
\item{dataset}{A CensusMapper dataset identifier.}

\item{level}{A geographic aggregation level for downloading data, e.g. CSD, CT, DA.}

\item{regions}{A json hash describing the geographic regions.}

\item{geo_format}{One of \code{"sf"} to return an \code{\link[sf]{sf}} object (the default; requires the \code{sf} package), \code{"sp"} to return a \code{\link[sp]{SpatialPolygonsDataFrame}} object (requires the \code{rgdal} package), or \code{NA} to append no geographical information to the returned data at all.}

\item{labels}{Set to "detailed" by default, but truncated Census variable names can be selected by setting labels = "short". Use cancensensus.labels() to return variable label information.}

\item{use_cache}{If set to TRUE (the default) data will be read from the local cache if available.}
}
\description{
Convenience function to load only census geography without data.
}
\examples{
census_data <- cancensus.load_geo(dataset='CA16', regions='{"CMA":["59933"]}', level='CSD')
}
\keyword{api}
\keyword{canada}
\keyword{census}
\keyword{data}
