% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_discovery.R
\name{explore_census_regions}
\alias{explore_census_regions}
\title{Interactively browse Census variables and regions on Censusmapper.ca in a new browser window}
\usage{
explore_census_regions(dataset = "CA16")
}
\arguments{
\item{dataset}{The dataset to query for available vectors, e.g. \code{'CA16'}.
Interactive tools available for the CA16, CA11, CA06, and CA01 Census datasets and
geographies.}
}
\description{
Finding the right Census variables or regions can be complicated.
\code{explore_census_vectors(dataset)} and \code{explore_census_regions(dataset)} will open a
new browser page or tab to an interactive Census variable and region exploration and selection
tool on the \href{https://censusmapper.ca/api}{Censusmapper.ca website}. Interactive
tools available for the CA16, CA11, CA06, and CA01 Census datasets and geographies.
}
\examples{
\dontrun{

explore_census_vectors(dataset = "CA16")

explore_census_regions(dataset = "CA11")

}
}
