% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_discovery.R
\name{explore_census_vectors}
\alias{explore_census_vectors}
\title{Interactively browse Census variable and region on Censusmapper.ca in a new browser window}
\usage{
explore_census_vectors()
}
\description{
Finding the right Census variables or regions can be complicated.
\code{explore_census_vectors()} will open a new browser page or tab to the interactive
Census variable and region tool built on the [Censusmapper.ca website](https://censusmapper.ca/api)
}
\examples{
\dontrun{

explore_census_vectors()
}
}
