% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{get_census}
\alias{get_census}
\alias{get_census_data}
\alias{get_census_geometry}
\title{Access to Canadian census data through the CensusMapper API}
\source{
Census data and boundary geometries are reproduced and distributed on
an "as is" basis with the permission of Statistics Canada (Statistics Canada
2006; 2011; 2016).
}
\usage{
get_census(dataset, level, regions, vectors = c(), geo_format = "sf",
  labels = "detailed", use_cache = TRUE,
  api_key = getOption("cancensus.api_key"))

get_census_data(dataset, level, regions, vectors, ...)

get_census_geometry(dataset, level, regions, geo_format = "sf", ...)
}
\arguments{
\item{dataset}{A CensusMapper dataset identifier.}

\item{level}{The census aggregation level to retrieve. One of \code{"Regions"}, \code{"PR"}, \code{"CMA"}, \code{"CD"}, \code{"CSD"}, \code{"CT"} or \code{"DA"}.}

\item{regions}{A named list of census regions to retrieve. Names must be valid census aggregation levels.}

\item{vectors}{An R vector containing the CensusMapper variable names of the census variables to download. If no vectors are specified only geographic data will get downloaded.}

\item{geo_format}{One of \code{"sf"} to return an \code{\link[sf]{sf}} object (the default; requires the \code{sf} package), \code{"sp"} to return a \code{\link[sp]{SpatialPolygonsDataFrame}} object (requires the \code{rgdal} package), or \code{NA} to append no geographical information to the returned data at all.}

\item{labels}{Set to "detailed" by default, but truncated Census variable names can be selected by setting labels = "short". Use cancensensus.labels() to return variable label information.}

\item{use_cache}{If set to TRUE (the default) data will be read from the local cache if available.}

\item{api_key}{An API key for the CensusMapper API. Defaults to \code{options()} and then the \code{CM_API_KEY} environment variable.}

\item{...}{Further arguments passed on to
\code{\link[cancensus]{get_census}}.}
}
\description{
These function allow convenient access to Canadian census data and boundary
files through the CensusMapper API. An API key is required to retrieve data.
}
\details{
\code{get_census_data} and \code{get_census_geometry} are convenience
functions that retrive only data or boundary information, respectively.

For help selecting regions and vectors, see \code{\link{list_census_regions}}
and \code{\link{list_census_vectors}}, or check out the interactive selection
tool at <https://censusmapper.ca/api>.
}
\examples{
# Query the API for data on dwellings in Vancouver, at the census subdivision
# level, and return the associated boundary files in `sf` format:
census_data <- get_census(dataset='CA16', regions=list(CMA="59933"),
                          vectors=c("v_CA16_408","v_CA16_409","v_CA16_410"),
                          level='CSD', geo_format = "sf")

# Make the same query, but this time drop descriptive vector names:
census_data <- get_census(dataset='CA16', regions=list(CMA="59933"),
                          vectors=c("v_CA16_408","v_CA16_409","v_CA16_410"),
                          level='CSD', geo_format = "sf", labels="short")

# Get details for truncated vectors:
census_vectors(census_data)

# Query the API for dwelling data in Vancouver at the census subdivision
# level, but not the associated boundary geometry.
vc_dwellings <- get_census_data(dataset='CA16', regions=list(CMA:"59933"),
                                vectors=c("v_CA16_408","v_CA16_409",
                                          "v_CA16_410"), level='CSD')

# Query the API for census subdivision boundary geometry within Vancouver.
vc_csds <- get_census_geometry(dataset='CA16', regions=list(CMA="59933"),
                               level='CSD', geo_format = "sf")

}
\keyword{api}
\keyword{canada}
\keyword{census}
\keyword{data}
