% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_geometry.R
\name{get_intersecting_geometries}
\alias{get_intersecting_geometries}
\title{Get identifiers for census regions interesecting a geometry}
\source{
Census data and boundary geographies are reproduced and distributed on
an "as is" basis with the permission of Statistics Canada (Statistics Canada
1996; 2001; 2006; 2011; 2016).
}
\usage{
get_intersecting_geometries(
  dataset,
  level,
  geometry,
  use_cache = TRUE,
  quiet = FALSE,
  api_key = getOption("cancensus.api_key")
)
}
\arguments{
\item{dataset}{A CensusMapper dataset identifier.}

\item{level}{The census aggregation level to retrieve. One of \code{"Regions"}, \code{"PR"}, \code{"CMA"}, \code{"CD"}, \code{"CSD"}, \code{"CT"}, \code{"DA"}, \code{"EA"} (for 1996 census), or \code{"DB"} (for 2001-2016 censuses)..}

\item{geometry}{An \code{sf} or \code{sfc} object}

\item{use_cache}{If set to TRUE (the default) data will be read from the local cache if available.}

\item{quiet}{When TRUE, suppress messages and warnings.}

\item{api_key}{An API key for the CensusMapper API. Defaults to \code{options()} and then the \code{CM_API_KEY} environment variable.}
}
\description{
This function returns a list of regions that intersect a give geometry. It is useful for example when
one is interested in census data for a particular geographic region that does not coincide with
well-known census geometries. The returned value can be used as the \code{regions} parameter in \code{get_census}
to get corresponding census geographies and variables that cover the give geometry.

This function requires advanced CensusMapper API access.
}
\examples{
# Query the API for the census tract contain the coordinates [-123.25149, 49.27026]
\dontrun{
point_geo <- sf::st_sfc(sf::st_point(c(-123.25149, 49.27026)),crs=4326)
regions <- get_intersecting_geometries(dataset = 'CA16', level = 'CT', geometry = point_geo)
census_data <- get_census(dataset='CA16', regions=regions,
                          vectors=c("v_CA16_408","v_CA16_409","v_CA16_410"),
                          level='CT')

}
}
