% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{list_census_vectors}
\alias{list_census_vectors}
\title{Query the CensusMapper API for available vectors for a given dataset.}
\usage{
list_census_vectors(dataset, use_cache = FALSE, quiet = TRUE)
}
\arguments{
\item{dataset}{The dataset to query for available vectors, e.g.
\code{"CA16"}.}

\item{use_cache}{If set to TRUE, data will be read from a local cache, if
available. If set to FALSE (the default), query the API for the data, and
refresh the local cache with the result.}

\item{quiet}{When FALSE, showss messages and warnings. Set to TRUE by default.}
}
\description{
Query the CensusMapper API for available vectors for a given dataset.
}
