% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{parent_census_vectors}
\alias{parent_census_vectors}
\title{List all parent variables from vector hierarchical based on a (sub-)list of census
variables returned by
\code{list_census_vectors} or \code{search_census_vectors}.}
\usage{
parent_census_vectors(vector_list)
}
\arguments{
\item{vector_list}{The list of vectors to be used}
}
\description{
List all parent variables from vector hierarchical based on a (sub-)list of census
variables returned by
\code{list_census_vectors} or \code{search_census_vectors}.
}
\examples{
\dontrun{
library(dplyr, warn.conflicts = FALSE)

list_census_vectors("CA16") \%>\%
  filter(vector == "v_CA16_4092") \%>\%
  parent_census_vectors()
}
}
