% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancensus.R
\name{parent_census_vectors}
\alias{parent_census_vectors}
\title{List all parent variables from vector hierarchichal based on a (sub-)list of census
variables returned by
\code{list_census_vectors} or \code{search_census_vectors}.}
\usage{
parent_census_vectors(base_list, dataset)
}
\arguments{
\item{dataset}{The dataset to query for available vectors, e.g.
\code{"CA16"}.}

\item{vector_list}{The list of vectors to be used}
}
\description{
List all parent variables from vector hierarchichal based on a (sub-)list of census
variables returned by
\code{list_census_vectors} or \code{search_census_vectors}.
}
\examples{
add_parent_census_vectors(vector_list, 'CA16')
}
