% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalled_data.R
\name{remove_recalled_chached_data}
\alias{remove_recalled_chached_data}
\title{Removes locally cached data that has been recalled}
\usage{
remove_recalled_chached_data()
}
\description{
Checks the local cached database for recalled data and remvoves cached data that has been recalled
}
\examples{
\dontrun{
remove_recalled_chached_data()
}
}
